ContabiliBotv0.4
Se creo un nuevo main para una interfaz nueva con funcionalidades distintas, las anteriores se guardaron en CBv.0.3 para reutilizar codigo.
