+++ .gitignore
@@ -1,0 +1,5 @@
+
+client_mod.xlsx
+client_mod.xlsx
+client.xlsx
+afip.xlsx
